+++ src/Endpoints/Suppression.php
@@ -7,7 +7,7 @@ use MailerSend\Common\Constants;
 use MailerSend\Common\HttpLayer;
 use MailerSend\Helpers\GeneralHelpers;
 
-abstract class Suppression extends AbstractEndpoint
+class Suppression extends AbstractEndpoint
 {
     protected string $endpoint;
 
@@ -47,7 +47,26 @@ abstract class Suppression extends AbstractEndpoint
         );
     }
 
-    abstract public function create(string $domainId, array $recipients);
+    /**
+     * @throws \JsonException
+     * @throws \MailerSend\Exceptions\MailerSendAssertException
+     * @throws \Psr\Http\Client\ClientExceptionInterface
+     */
+    public function create(string $domainId, array $recipients): array
+    {
+        GeneralHelpers::assert(
+            fn () => Assertion::minLength($domainId, 1, 'Domain id is required.')
+                && Assertion::notEmpty($recipients, 'Recipients is required.')
+        );
+
+        return $this->httpLayer->post(
+            $this->buildUri($this->endpoint),
+            [
+                'domain_id' => $domainId,
+                'recipients' => $recipients,
+            ]
+        );
+    }
 
     /**
      * @throws \JsonException
